[clang] [diagnostics] Add `-fdiagnostics-add-output` switch for SARIF
This change introduces the new `-fdiagnostics-add-output=` switch, which adds an additional output file for diagnostics. The syntax of the switch, which matches an existing GCC switch, is `-fdiagnostics-add-output=format:key1=value1,key2=value2...`. `format` specifies the format of the diagnostics; currently, only `sarif` is supported, but it would be easy enough to add other formats. The optional key-value pairs specify options specific to that format. For `sarif`, the valid options are:

- `file=<path>` - Specifies the path to the generated SARIF file. Unlike GCC, this property is required. We could make it optional by providing a default based on the input filename, but plumbing that through would have made the implementation more complex.
- `version=<version>` - Specifies the SARIF schema version. Only `2.1` is currently supported.

The `-fdiagnostics-add-output=` switch can be specified multiple times. Each occurrence adds an additional output sink, so multiple SARIF logs are possible with different file paths.
- I've added a positive test for the new switch, including specifying it multiple times.
- I've added a negative test for various malformed format/key/value strings.
- I've updated our existing SARIF test to use the same normalize_sarif filter that the Static Analyzer's SARIF tests use, rather than a complicated set of CHECK directives.

rdar://160939516